refractor: Refactor simulation class structure
