+++ src/engine/__init__.py
@@ -1,1 +1,1 @@
-from .simulation import Simulation
+from .Simulation import Simulation
