+++ src/engine/utils/__init__.py
@@ -1,0 +1,2 @@
+from .citizens_utils import *
+from .connectivity_matrix_utils import *
