+++ src/engine/utils/citizens_utils.py
@@ -1,0 +1,13 @@
+# ================================================================
+# 0. Section: Imports
+# ================================================================
+import numpy as np
+
+
+
+# ================================================================
+# 1. Section: Iterative Updates
+# ================================================================
+def age_citizens(citizens: np.ndarray) -> None:
+    for citizen in citizens: 
+        if(citizen.state == "alive"): citizen.age_citizen()
